Read from database2

+++ app/src/main/java/com/example/onegmall/DonutCurrencyCheck.java
@@ -64,7 +64,7 @@ public class DonutCurrencyCheck extends AppCompatActivity {
     private void showData(DataSnapshot dataSnapshot) {
         for(DataSnapshot ds : dataSnapshot.getChildren()){
             UserInfo uData = new UserInfo();
-            Double c=(Double)ds.child("currency").getValue();
+            Long c=(Long)ds.child("currency").getValue();
             //uData.setCurrency(ds.child(userID).getValue(UserInfo.class).getCurrency());
             //uData.setLifeAfter(ds.child(userID).getValue(UserInfo.class).getLifeAfter());
